passando responsabilidade de busca e download de imagens para um arquivo separado
